Changed the script infra
