+++ Assets/Scripts/Spawner.cs
@@ -7,10 +7,9 @@ public class Spawner : MonoBehaviour
     private MeshRenderer mr;
     private bool glow, placedUnit;
     public static bool overlay;
+    private Transform _transform;
 
     private string spawnerTag = "Spawner";
-
-    private Transform _transform;
 
     //Gets the mesh of the spawners so they can be changed.
     private void Awake()
