+++ Assets/Scripts/UnitConfig.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f5a5c9420dc52d546aae38051747867a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
